# api/openapi.yaml
openapi: 3.0.0
info:
  title: EliteGate Examples Mock Services API
  description: Mock backend services for API Gateway testing
  version: 1.0.0
  contact:
    name: API Support

servers:
  - url: http://localhost:9001
    description: User Service
  - url: http://localhost:9002
    description: Order Service

tags:
  - name: Health
    description: Health check endpoints
  - name: Users
    description: User management operations
  - name: Orders
    description: Order management operations
  - name: Debug
    description: Debug and diagnostic endpoints

paths:
  /health:
    get:
      summary: Health check
      operationId: getHealth
      tags:
        - Health
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
      x-servers:
        - url: http://localhost:9001
        - url: http://localhost:9002

  /users:
    get:
      summary: List all users
      operationId: listUsers
      tags:
        - Users
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserList'
      x-servers:
        - url: http://localhost:9001

    post:
      summary: Create a new user
      operationId: createUser
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-servers:
        - url: http://localhost:9001

  /users/{id}:
    get:
      summary: Get user by ID
      operationId: getUserById
      tags:
        - Users
      parameters:
        - name: id
          in: path
          required: true
          description: User ID
          schema:
            type: integer
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid user ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-servers:
        - url: http://localhost:9001

  /orders:
    get:
      summary: List all orders
      operationId: listOrders
      tags:
        - Orders
      responses:
        '200':
          description: List of orders
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderList'
      x-servers:
        - url: http://localhost:9002

    post:
      summary: Create a new order
      operationId: createOrder
      tags:
        - Orders
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Order created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-servers:
        - url: http://localhost:9002

  /orders/{id}:
    get:
      summary: Get order by ID
      operationId: getOrderById
      tags:
        - Orders
      parameters:
        - name: id
          in: path
          required: true
          description: Order ID
          schema:
            type: integer
      responses:
        '200':
          description: Order found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Invalid order ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-servers:
        - url: http://localhost:9002

  /debug:
    get:
      summary: Echo request details for debugging
      operationId: debugRequest
      tags:
        - Debug
      parameters:
        - name: test_param
          in: query
          description: Optional query parameter for testing
          schema:
            type: string
        - name: X-Custom-Header
          in: header
          description: Custom header for testing header propagation
          schema:
            type: string
      responses:
        '200':
          description: Debug information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DebugResponse'
      x-servers:
        - url: http://localhost:9001
        - url: http://localhost:9002

components:
  schemas:
    HealthResponse:
      type: object
      required:
        - status
        - service
      properties:
        status:
          type: string
          example: ok
        service:
          type: string
          example: user-service

    User:
      type: object
      required:
        - id
        - name
        - email
        - age
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Alice Johnson
        email:
          type: string
          format: email
          example: alice@example.com
        age:
          type: integer
          example: 28

    CreateUserRequest:
      type: object
      required:
        - name
        - email
        - age
      properties:
        name:
          type: string
          example: John Doe
        email:
          type: string
          format: email
          example: john@example.com
        age:
          type: integer
          example: 30

    UserList:
      type: object
      required:
        - users
        - count
      properties:
        users:
          type: array
          items:
            $ref: '#/components/schemas/User'
        count:
          type: integer
          example: 3

    Order:
      type: object
      required:
        - id
        - user_id
        - product
        - quantity
        - price
        - status
      properties:
        id:
          type: integer
          example: 1
        user_id:
          type: integer
          example: 1
        product:
          type: string
          example: Laptop
        quantity:
          type: integer
          example: 1
        price:
          type: number
          format: double
          example: 999.99
        status:
          type: string
          enum:
            - pending
            - shipped
            - delivered
            - cancelled
          example: shipped

    CreateOrderRequest:
      type: object
      required:
        - user_id
        - product
        - quantity
        - price
      properties:
        user_id:
          type: integer
          example: 1
        product:
          type: string
          example: Mouse
        quantity:
          type: integer
          example: 2
        price:
          type: number
          format: double
          example: 29.99

    OrderList:
      type: object
      required:
        - orders
        - count
      properties:
        orders:
          type: array
          items:
            $ref: '#/components/schemas/Order'
        count:
          type: integer
          example: 3

    DebugResponse:
      type: object
      properties:
        method:
          type: string
          example: GET
        path:
          type: string
          example: /debug
        headers:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          example:
            User-Agent: ["curl/7.64.1"]
            Accept: ["*/*"]
            X-Custom-Header: ["test-value"]
        query_params:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          example:
            test_param: ["value"]
        remote_addr:
          type: string
          example: "172.17.0.1:54321"
        timestamp:
          type: string
          format: date-time
          example: "2024-01-15T10:30:45Z"

    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: "user not found"